chore: gather monitor configs into configs.monitors
