+++ src/prime_rl/monitors/__init__.py
@@ -12,7 +12,7 @@ import os
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
-from prime_rl.configs.shared import FileMonitorConfig, PrimeMonitorConfig, WandbConfig
+from prime_rl.configs.monitors import FileMonitorConfig, PrimeMonitorConfig, WandbConfig
 from prime_rl.monitors.base import Monitor
 from prime_rl.monitors.file import FileMonitor
 from prime_rl.monitors.prime import PrimeMonitor
